requirements.txt
addict==2.2.1
PyYAML==5.3.1
click==7.0
pre-commit==1.18.3
pandas==1.0.3
yfinance==0.1.54
yahoofinancials==1.5
streamlit==0.67.1
loguru==0.2.5